fix(whatsapp): keep setup auth checks lightweight
