+++ extensions/whatsapp/src/auth-store.ts
@@ -6,6 +6,12 @@ import { DEFAULT_ACCOUNT_ID } from "openclaw/plugin-sdk/routing";
 import { info, success } from "openclaw/plugin-sdk/runtime-env";
 import { getChildLogger } from "openclaw/plugin-sdk/runtime-env";
 import { defaultRuntime, type RuntimeEnv } from "openclaw/plugin-sdk/runtime-env";
+import {
+  readWebAuthStateDetails,
+  waitForWebAuthBarrier,
+  webAuthExistsAt,
+  type WhatsAppWebAuthState,
+} from "./auth-state.js";
 import { resolveOAuthDir } from "./auth-store.runtime.js";
 import {
   assertWebCredsPathRegularFileOrMissing,
@@ -16,20 +22,15 @@ import {
   resolveWebCredsPath,
   statWebCredsFileSync,
 } from "./creds-files.js";
-import {
-  waitForCredsSaveQueueWithTimeout,
-  writeWebCredsRawAtomically,
-  type CredsQueueWaitResult,
-} from "./creds-persistence.js";
+import { writeWebCredsRawAtomically } from "./creds-persistence.js";
 import { resolveComparableIdentity, type WhatsAppSelfIdentity } from "./identity.js";
 import { resolveUserPath, type WebChannel } from "./text-runtime.js";
 export { hasWebCredsSync, resolveWebCredsBackupPath, resolveWebCredsPath };
+export { formatWhatsAppWebAuthStatusState, type WhatsAppWebAuthState } from "./auth-state.js";
 
 export const WHATSAPP_AUTH_UNSTABLE_CODE = "whatsapp-auth-unstable";
 
-const authStoreLogger = getChildLogger({ module: "web-auth-store" });
 const emptyWebSelfId = () => ({ e164: null, jid: null, lid: null }) as const;
-export type WhatsAppWebAuthState = "linked" | "not-linked" | "unstable";
 
 export class WhatsAppAuthUnstableError extends Error {
   readonly code = WHATSAPP_AUTH_UNSTABLE_CODE;
@@ -48,23 +49,6 @@ export const WA_WEB_AUTH_DIR = resolveDefaultWebAuthDir();
 
 export function readCredsJsonRaw(filePath: string): string | null {
   return readWebCredsJsonRawSync(filePath);
-}
-
-async function waitForWebAuthBarrier(
-  authDir: string,
-  context: string,
-): Promise<CredsQueueWaitResult> {
-  const result = await waitForCredsSaveQueueWithTimeout(authDir);
-  if (result === "timed_out") {
-    authStoreLogger.warn(
-      {
-        authDir,
-        context,
-      },
-      "timed out waiting for queued WhatsApp creds save before auth read",
-    );
-  }
-  return result;
 }
 
 function isValidJson(raw: string): boolean {
@@ -128,65 +112,17 @@ export async function restoreCredsFromBackupIfNeeded(
 }
 
 export async function webAuthExists(authDir: string = resolveDefaultWebAuthDir()) {
-  const resolvedAuthDir = resolveUserPath(authDir);
-  const credsPath = resolveWebCredsPath(resolvedAuthDir);
-  const raw = await readWebCredsJsonRaw(credsPath);
-  if (!raw) {
-    return false;
-  }
-  try {
-    JSON.parse(raw);
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-function resolveWebAuthState(params: {
-  linked: boolean;
-  barrierResult: CredsQueueWaitResult;
-}): WhatsAppWebAuthState {
-  if (params.barrierResult === "timed_out") {
-    return "unstable";
-  }
-  return params.linked ? "linked" : "not-linked";
-}
-
-async function readWebAuthStateCore(
-  authDir: string,
-  context: string,
-): Promise<{ authDir: string; linked: boolean; state: WhatsAppWebAuthState }> {
-  const resolvedAuthDir = resolveUserPath(authDir);
-  const barrierResult = await waitForWebAuthBarrier(resolvedAuthDir, context);
-  const linked = await webAuthExists(resolvedAuthDir);
-  return {
-    authDir: resolvedAuthDir,
-    linked,
-    state: resolveWebAuthState({ linked, barrierResult }),
-  };
-}
-
-export function formatWhatsAppWebAuthStatusState(state: WhatsAppWebAuthState): string {
-  switch (state) {
-    case "linked":
-      return "linked";
-    case "not-linked":
-      return "not linked";
-    case "unstable":
-      return "auth stabilizing";
-  }
-  const exhaustive: never = state;
-  return exhaustive;
+  return await webAuthExistsAt(authDir);
 }
 
 export async function readWebAuthState(
   authDir: string = resolveDefaultWebAuthDir(),
 ): Promise<WhatsAppWebAuthState> {
-  return (await readWebAuthStateCore(authDir, "readWebAuthState")).state;
+  return (await readWebAuthStateDetails(authDir, "readWebAuthState")).state;
 }
 
 export async function readWebAuthSnapshot(authDir: string = resolveDefaultWebAuthDir()) {
-  const auth = await readWebAuthStateCore(authDir, "readWebAuthSnapshot");
+  const auth = await readWebAuthStateDetails(authDir, "readWebAuthSnapshot");
   return {
     state: auth.state,
     authAgeMs: auth.state === "linked" ? getWebAuthAgeMs(auth.authDir) : null,
