+++ extensions/whatsapp/src/channel.setup.ts
@@ -1,7 +1,8 @@
 // Whatsapp plugin module implements channel.setup behavior.
 import type { ChannelPlugin } from "openclaw/plugin-sdk/core";
 import type { ResolvedWhatsAppAccount } from "./accounts.js";
-import { isWhatsAppAuthConfigured } from "./channel-runtime-loader.js";
+import { formatWhatsAppConfigAllowFromEntries } from "./allowlist-format.js";
+import { readWebAuthState } from "./auth-state.js";
 import {
   resolveWhatsAppGroupRequireMention,
   resolveWhatsAppGroupToolPolicy,
@@ -9,6 +10,10 @@ import {
 import { whatsappSetupAdapter } from "./setup-core.js";
 import { createWhatsAppPluginBase, whatsappSetupWizardProxy } from "./shared.js";
 import { detectWhatsAppLegacyStateMigrations } from "./state-migrations.js";
+
+async function isWhatsAppAuthConfigured(account: ResolvedWhatsAppAccount): Promise<boolean> {
+  return (await readWebAuthState(account.authDir)) === "linked";
+}
 
 export const whatsappSetupPlugin: ChannelPlugin<ResolvedWhatsAppAccount> = {
   ...createWhatsAppPluginBase({
@@ -18,7 +23,8 @@ export const whatsappSetupPlugin: ChannelPlugin<ResolvedWhatsAppAccount> = {
     },
     setupWizard: whatsappSetupWizardProxy,
     setup: whatsappSetupAdapter,
-    isConfigured: async (account) => await isWhatsAppAuthConfigured(account.authDir),
+    formatAllowFrom: formatWhatsAppConfigAllowFromEntries,
+    isConfigured: isWhatsAppAuthConfigured,
   }),
   lifecycle: {
     detectLegacyStateMigrations: ({ oauthDir }) =>
