+++ extensions/whatsapp/src/config-accessors.ts
@@ -1,17 +1,13 @@
 // Whatsapp helper module supports config accessors behavior.
 import type { OpenClawConfig } from "openclaw/plugin-sdk/config-contracts";
 import { resolveWhatsAppAccount } from "./accounts.js";
-import { normalizeWhatsAppAllowFromEntries } from "./normalize-target.js";
+export { formatWhatsAppConfigAllowFromEntries } from "./allowlist-format.js";
 
 export function resolveWhatsAppConfigAllowFrom(params: {
   cfg: OpenClawConfig;
   accountId?: string | null;
 }): string[] {
   return [...(resolveWhatsAppAccount(params).allowFrom ?? [])];
-}
-
-export function formatWhatsAppConfigAllowFromEntries(allowFrom: Array<string | number>): string[] {
-  return normalizeWhatsAppAllowFromEntries(allowFrom);
 }
 
 export function resolveWhatsAppConfigDefaultTo(params: {
