+++ extensions/whatsapp/src/setup-surface.ts
@@ -7,7 +7,7 @@ import {
   type OpenClawConfig,
 } from "openclaw/plugin-sdk/setup";
 import { listWhatsAppAccountIds, resolveWhatsAppAuthDir } from "./accounts.js";
-import { formatWhatsAppWebAuthStatusState, readWebAuthState } from "./auth-store.js";
+import { formatWhatsAppWebAuthStatusState, readWebAuthState } from "./auth-state.js";
 
 const t = createSetupTranslator();
